use one singletone instance

+++ Assets/Scripts/Utils.cs
@@ -9,7 +9,7 @@ public static class Utils
 
     public static float GetTupleRatio((int, int) tuple)
     {
-        return (float)tuple.Item1 / (float)tuple.Item2;
+        return (float)tuple.Item1 / tuple.Item2;
     }
 
     public static float GetAverage(params float[] prs)
